pyproject.toml
[tool.poetry]
name = "answerbook-api"
version = "0.1.0"
description = ""
authors = ["Ivan Procaccini <ivanprocaccini905@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
typer = "^0.12.3"
freezegun = "^1.5.1"
pydantic-settings = "^2.3.3"
sqlmodel = "^0.0.19"
sqlalchemy-utils = "^0.41.2"
types-pyyaml = "^6.0.12.20240311"
alembic = "^1.13.1"
requests = "^2.32.3"
types-requests = "^2.32.0.20240622"
langchain = "^0.2.6"
langchain-openai = "^0.1.14"
authlib = "^1.3.1"
passlib = "^1.7.4"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"
python-ldap = "^3.4.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
pytest-factoryboy = "^2.7.0"
mypy = "^1.10.0"
black = "^24.4.2"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
exclude = "venv"
disable_error_code = "var-annotated"
ignore_missing_imports = true
